+++ build.gradle
@@ -67,16 +67,8 @@ repositories {
 }
 
 dependencies {
-	implementation('de.unistuttgart.iste.gits:gits-common') {
-		version {
-			branch = 'main'
-		}
-	}
-	implementation('de.unistuttgart.iste.gits:content_service') { // include content service for DTOs
-		version {
-			branch = 'main'
-		}
-	}
+	implementation 'de.unistuttgart.iste.gits:gits-common:0.1.1'
+	implementation 'de.unistuttgart.iste.gits:content_service:0.1.2'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-graphql'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
