+++ src/main/java/de/unistuttgart/iste/gits/skilllevel_service/controller/SubscriptionController.java
@@ -20,6 +20,9 @@ import reactor.core.publisher.Mono;
 import java.util.Map;
 import java.util.UUID;
 
+/**
+ * Controller for Dapr pubsub topic subscriptions.
+ */
 @RestController
 @RequiredArgsConstructor
 @Slf4j
@@ -27,28 +30,35 @@ public class SubscriptionController {
     private final SkillLevelService skillLevelService;
     private final ContentServiceClient contentServiceClient;
 
+    /**
+     * Dapr topic subscription to recalculate the skill levels of a user for a specific chapter when the user
+     * completes an assessment in that chapter.
+     */
     @Topic(name = "user-progress-updated", pubsubName = "gits")
     @PostMapping(path = "/skilllevel-service/user-progress-pubsub")
-    public Mono<SkillLevels> onUserProgress(@RequestBody CloudEvent<UserProgressLogEvent> cloudEvent,
-                                            @RequestHeader Map<String, String> headers) {
+    public Mono<Void> onUserProgress(@RequestBody CloudEvent<UserProgressLogEvent> cloudEvent) {
         log.info("Received event: {}", cloudEvent.getData());
 
-        return Mono.fromCallable(() -> {
+        return Mono.fromRunnable(() -> {
             try {
                 UUID chapterId =
                         contentServiceClient.getChapterIdOfContent(cloudEvent.getData().getContentId());
-                return skillLevelService.recalculateLevels(chapterId, cloudEvent.getData().getUserId());
+                skillLevelService.recalculateLevels(chapterId, cloudEvent.getData().getUserId());
             } catch (Exception e) {
+                // we need to catch all exceptions because otherwise if some invalid data is in the message queue
+                // it will never get processed and instead the service will just crash forever
                 log.error("Error while processing user progress event", e);
-                return null;
             }
         });
     }
 
+    /**
+     * Dapr topic subscription to delete the stored skill level data of users for a specific chapter when the chapter
+     * is deleted.
+     */
     @Topic(name = "chapter-changes", pubsubName = "gits")
     @PostMapping(path = "/skilllevel-service/chapter-changes-pubsub")
-    public Mono<Void> onChapterChanged(@RequestBody CloudEvent<ChapterChangeEvent> cloudEvent,
-                                      @RequestHeader Map<String, String> headers) {
+    public Mono<Void> onChapterChanged(@RequestBody CloudEvent<ChapterChangeEvent> cloudEvent) {
         return Mono.fromRunnable(() -> {
             try {
                 if(cloudEvent.getData().getOperation() != CrudOperation.DELETE)
@@ -58,6 +68,8 @@ public class SubscriptionController {
                     skillLevelService.deleteSkillLevelsForChapter(chapterId);
                 }
             } catch (Exception e) {
+                // we need to catch all exceptions because otherwise if some invalid data is in the message queue
+                // it will never get processed and instead the service will just crash forever
                 log.error("Error while processing course change event", e);
             }
         });
