+++ src/main/resources/application-prod.properties
@@ -8,6 +8,9 @@ spring.sql.init.mode=always
 spring.jpa.show-sql=true
 spring.sql.init.continue-on-error=true
 spring.jpa.hibernate.ddl-auto=update
+
+# URL to the GraphQL endpoint of the content service. The skilllevel service uses this to fetch information
+# about assignments it needs to calculate the users' skill levels
 content_service.url=http://app-content:4001/graphql
