+++ src/main/resources/graphql/common/scalars.graphqls
