+++ src/main/resources/graphql/service/schema.graphqls
